endpoints para crear y logear clientes

+++ models/product.js
@@ -18,7 +18,7 @@ module.exports = (sequelize, DataTypes) => {
   }
   Product.init({
     name: DataTypes.STRING,
-    price: DataTypes.FLOAT,
+    price: DataTypes.NUMERIC(10,2),
     CategoryId: DataTypes.INTEGER
   }, {
     sequelize,
